Fix compiler warnings

+++ changes/change.cpp
@@ -142,7 +142,7 @@ string changes_change (void * webserver_request)
     int timestap = database_notes.get_modified (note);
     timestamps.push_back (timestap);
   }
-  quick_sort (timestamps, notes, 0, notes.size ());
+  quick_sort (timestamps, notes, 0, static_cast<int>(notes.size ()));
   reverse (notes.begin(), notes.end());
